split out config module
