+++ lib/remote.rb
@@ -1,0 +1,7 @@
+module Config
+  module Drivers
+    #Class for configuring remote execution
+    class Remote
+    end
+  end
+end
